requirements.txt
langchain
langchain-community
openai
tiktoken
requests
python-dotenv
langchain-openai